Refactor CourseDetailViewModel Rating Logic
Moved logic to fetch `User` and `RatingSummary` into use cases to keep `CourseDetailViewModel` lightweight.
Modified `CourseRatingUtils.showRating` and `BaseContainerFragment.setRatings` to utilize `RatingSummary` object instead of extracting via generic JSON, making typing stronger.
Refactored unit tests to cover the use cases and mock the newly passed properties.
