+++ app/src/main/java/org/ole/planet/myplanet/ui/courses/RatingSummaryProvider.kt
@@ -1,10 +1,10 @@
-package org.ole.planet.myplanet.domain.usecase
+package org.ole.planet.myplanet.ui.courses
 
 import javax.inject.Inject
 import kotlinx.coroutines.withContext
 import org.ole.planet.myplanet.model.RealmUser
-import org.ole.planet.myplanet.repository.RatingsRepository
 import org.ole.planet.myplanet.repository.RatingSummary
+import org.ole.planet.myplanet.repository.RatingsRepository
 import org.ole.planet.myplanet.services.UserSessionManager
 import org.ole.planet.myplanet.utils.DispatcherProvider
 
@@ -13,7 +13,7 @@ data class RatingSummaryModel(
     val ratingSummary: RatingSummary?
 )
 
-class GetRatingSummaryUseCase @Inject constructor(
+class RatingSummaryProvider @Inject constructor(
     private val ratingsRepository: RatingsRepository,
     private val userSessionManager: UserSessionManager,
     private val dispatcherProvider: DispatcherProvider
